# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/encounters": {
      "get": {
        "operationId": "EncountersController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EncounterResponseDto"
                          }
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      },
      "post": {
        "operationId": "EncountersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEncounterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/EncounterResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
...
    "/api/v1/encounters/{encounterId}": {
      "get": {
        "operationId": "EncountersController_detail",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/EncounterResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/events": {
      "get": {
        "operationId": "EncountersController_events",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/EncounterEventResponseDto"
                          }
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/transitions": {
      "post": {
        "operationId": "EncountersController_transition",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionEncounterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/EncounterResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/closures": {
      "post": {
        "operationId": "EncountersController_close",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseEncounterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/EncounterResponseDto"
                        },
                        "meta": {
                          "type": "object"
                        },
                        "requestId": {
                          "type": "string",
                          "format": "uuid"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/transition": {
      "post": {
        "operationId": "EncountersController_transitionAlias",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionEncounterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
    "/api/v1/encounters/{encounterId}/close": {
      "post": {
        "operationId": "EncountersController_closeAlias",
        "parameters": [
          {
            "name": "encounterId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseEncounterRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "encounters"
        ]
      }
    },
...
    "/api/v1/audit/encounters/{id}": {
      "get": {
        "operationId": "AuditController_encounter",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "DermaHealth Backend API",
    "description": "Clinic workflow backend: identity, scheduling, clinical, workflow, EMR, CRM.",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      }
    },
    "schemas": {
...
      "EncounterResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "parentEncounterId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "appointmentId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "type": {
            "type": "string",
            "enum": [
              "standard",
              "emergency",
              "follow_up",
              "remote"
            ]
          },
          "origin": {
            "type": "string",
            "enum": [
              "appointment",
              "walk_in",
              "follow_up_request"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "registered",
              "intake_in_progress",
              "intake_complete",
              "ai_assessed",
              "routed",
              "checked_in",
              "under_doctor_review",
              "awaiting_results",
              "diagnosed",
              "plan_approved",
              "workflow_active",
              "in_progress",
              "results_complete",
              "final_review",
              "discharge_ready",
              "record_signed",
              "closed",
              "post_visit_monitoring",
              "escalated",
              "follow_up_linked"
            ]
          },
          "department": {
            "type": "string"
          },
          "room": {
            "type": "object",
            "nullable": true
          },
          "queueNumber": {
            "type": "object",
            "nullable": true
          },
          "peopleAheadInQueue": {
            "type": "object",
            "nullable": true
          },
          "estimatedWaitMinutes": {
            "type": "object",
            "nullable": true
          },
          "currentDoctorId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "blockingCondition": {
            "type": "object",
            "nullable": true
          },
          "version": {
            "type": "number"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "patientId",
          "parentEncounterId",
          "appointmentId",
          "type",
          "origin",
          "status",
          "department",
          "room",
          "queueNumber",
          "peopleAheadInQueue",
          "estimatedWaitMinutes",
          "currentDoctorId",
          "blockingCondition",
          "version",
          "createdAt",
          "updatedAt"
        ]
      },
      "EncounterEventResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "at": {
            "type": "string",
            "format": "date-time"
          },
          "label": {
            "type": "string"
          },
          "kind": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "success",
              "danger"
            ]
          }
        },
        "required": [
          "id",
          "at",
          "label",
          "kind"
        ]
      },
      "CreateEncounterRequest": {
        "type": "object",
        "properties": {
          "patientId": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": [
              "standard",
              "emergency",
              "follow_up",
              "remote"
            ]
          },
          "origin": {
            "type": "string",
            "enum": [
              "walk_in",
              "follow_up_request"
            ]
          },
          "department": {
            "type": "string"
          },
          "parentEncounterId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "patientId",
          "type",
          "origin",
          "department"
        ]
      },
      "TransitionEncounterRequest": {
        "type": "object",
        "properties": {
          "toStatus": {
            "type": "string",
            "enum": [
              "registered",
              "intake_in_progress",
              "intake_complete",
              "ai_assessed",
              "routed",
              "checked_in",
              "under_doctor_review",
              "awaiting_results",
              "diagnosed",
              "plan_approved",
              "workflow_active",
              "in_progress",
              "results_complete",
              "final_review",
              "discharge_ready",
              "record_signed",
              "closed",
              "post_visit_monitoring",
              "escalated",
              "follow_up_linked"
            ]
          },
          "reason": {
            "type": "string"
          },
          "blockingCondition": {
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "toStatus",
          "version"
        ]
      },
      "CloseEncounterRequest": {
        "type": "object",
        "properties": {
          "version": {
            "type": "number"
          }
        },
        "required": [
          "version"
        ]
      },
...
    }
  }
}